Moved some methods in organizer database class to the User database class
